+++ src/components/page/update/sections/location-section.tsx
@@ -1,6 +1,17 @@
+import {
+  Card,
+  CardContent,
+  CardHeader,
+  CardTitle,
+} from "@/src/components/ui/card";
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/src/components/ui/form";
 import { Input } from "@/src/components/ui/input";
-import { Label } from "@/src/components/ui/label";
-import { Card, CardContent, CardHeader, CardTitle } from "@/src/components/ui/card";
 import {
   Select,
   SelectContent,
@@ -8,110 +19,114 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/src/components/ui/select";
+import { TOrganizationSchema } from "@/src/schemas/organization";
 import { canadianProvinces } from "@/src/types/auth-organizations-types";
+import { UseFormReturn } from "react-hook-form";
 
 interface LocationSectionProps {
-  formData: {
-    country: string;
-    state: string;
-    city: string;
-    address: string;
-  };
-  errors: Record<string, string[]>;
-  onInputChange: (field: string) => (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => void;
-  onSelectChange: (field: string) => (value: string) => void;
+  form: UseFormReturn<TOrganizationSchema>;
 }
 
-export function LocationSection({
-  formData,
-  errors,
-  onInputChange,
-  onSelectChange,
-}: LocationSectionProps) {
+export function LocationSection({ form }: LocationSectionProps) {
   return (
     <Card>
       <CardHeader>
         <CardTitle className="text-xl font-semibold">Location</CardTitle>
       </CardHeader>
       <CardContent className="space-y-6">
-        <div className="space-y-3">
-          <Label htmlFor="country" className="text-sm font-semibold text-gray-700 dark:text-gray-300">
-            Country
-          </Label>
-          <Input
-            id="country"
-            name="country"
-            type="text"
-            value={formData.country}
-            readOnly
-            className="bg-gray-100 dark:bg-gray-800"
-          />
-        </div>
-
-        <div className="space-y-3">
-          <Label htmlFor="state" className="text-sm font-semibold text-gray-700 dark:text-gray-300">
-            Province/Territory *
-          </Label>
-          <div className="w-full max-w-md">
-            <Select
-              name="state"
-              required
-              value={formData.state}
-              onValueChange={onSelectChange("state")}
-            >
-              <SelectTrigger className={`w-full ${errors?.state ? "border-red-500" : ""}`}>
-                <SelectValue placeholder="Select your province or territory" />
-              </SelectTrigger>
-              <SelectContent>
-                {canadianProvinces.map((province) => (
-                  <SelectItem key={province.value} value={province.value}>
-                    {province.label}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-          {errors?.state && (
-            <p className="text-sm text-red-500">{errors.state[0]}</p>
+        <FormField
+          control={form.control}
+          name="country"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Country
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="text"
+                  readOnly
+                  className="bg-gray-100 dark:bg-gray-800"
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
           )}
-        </div>
+        />
 
-        <div className="space-y-3">
-          <Label htmlFor="city" className="text-sm font-semibold text-gray-700 dark:text-gray-300">
-            City *
-          </Label>
-          <Input
-            id="city"
-            name="city"
-            type="text"
-            placeholder="Toronto"
-            required
-            value={formData.city}
-            onChange={onInputChange("city")}
-            className={errors?.city ? "border-red-500" : ""}
-          />
-          {errors?.city && (
-            <p className="text-sm text-red-500">{errors.city[0]}</p>
+        <FormField
+          control={form.control}
+          name="state"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Province/Territory *
+              </FormLabel>
+              <Select onValueChange={field.onChange} defaultValue={field.value}>
+                <FormControl>
+                  <SelectTrigger className="w-full text-black">
+                    <SelectValue placeholder="Select your province or territory" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent>
+                  {canadianProvinces.map((province) => (
+                    <SelectItem key={province.value} value={province.value}>
+                      {province.label}
+                    </SelectItem>
+                  ))}
+                </SelectContent>
+              </Select>
+              <FormMessage />
+            </FormItem>
           )}
-        </div>
+        />
 
-        <div className="space-y-3">
-          <Label htmlFor="address" className="text-sm font-semibold text-gray-700 dark:text-gray-300">
-            Street Address *
-          </Label>
-          <Input
-            id="address"
-            name="address"
-            placeholder="123 Main Street"
-            required
-            value={formData.address}
-            onChange={onInputChange("address")}
-            className={errors?.address ? "border-red-500" : ""}
-          />
-          {errors?.address && (
-            <p className="text-sm text-red-500">{errors.address[0]}</p>
+        <FormField
+          control={form.control}
+          name="city"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                City *
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="text"
+                  placeholder="Montreal"
+                  className={
+                    form.formState.errors?.city ? "border-red-500" : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
           )}
-        </div>
+        />
+
+        <FormField
+          control={form.control}
+          name="address"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Street Address *
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="text"
+                  placeholder="123 Main Street"
+                  className={
+                    form.formState.errors?.address ? "border-red-500" : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
       </CardContent>
     </Card>
   );
