+++ src/components/page/update/sections/mission-projects-section.tsx
@@ -1,54 +1,72 @@
+import {
+  Card,
+  CardContent,
+  CardHeader,
+  CardTitle,
+} from "@/src/components/ui/card";
+import { Checkbox } from "@/src/components/ui/checkbox";
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/src/components/ui/form";
 import { Input } from "@/src/components/ui/input";
 import { Label } from "@/src/components/ui/label";
 import { Textarea } from "@/src/components/ui/textarea";
-import { Checkbox } from "@/src/components/ui/checkbox";
-import { Card, CardContent, CardHeader, CardTitle } from "@/src/components/ui/card";
 import { projectAreaOptions } from "@/src/types/auth-organizations-types";
 
+import { UseFormReturn } from "react-hook-form";
+
+import { TOrganizationSchema } from "@/src/schemas/organization";
+
 interface MissionProjectsSectionProps {
-  formData: {
-    missionStatement: string;
-    projectAreas: string[];
-    websiteUrl: string;
-    facebookUrl: string;
-    twitterUrl: string;
-    instagramUrl: string;
-    linkedinUrl: string;
-  };
-  errors: Record<string, string[]>;
-  onInputChange: (field: string) => (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => void;
-  onProjectAreaChange: (area: string, checked: boolean) => void;
+  form: UseFormReturn<TOrganizationSchema>;
 }
 
-export function MissionProjectsSection({
-  formData,
-  errors,
-  onInputChange,
-  onProjectAreaChange,
-}: MissionProjectsSectionProps) {
+export function MissionProjectsSection({ form }: MissionProjectsSectionProps) {
+  const onProjectAreaChange = (area: string, checked: boolean) => {
+    const current = form.getValues("projectAreas") || [];
+    const updated = checked
+      ? [...current, area]
+      : current.filter((a) => a !== area);
+    form.setValue("projectAreas", updated, {
+      shouldValidate: true,
+      shouldDirty: true,
+    });
+  };
   return (
     <Card>
       <CardHeader>
-        <CardTitle className="text-xl font-semibold">Mission & Projects</CardTitle>
+        <CardTitle className="text-xl font-semibold">
+          Mission & Projects
+        </CardTitle>
       </CardHeader>
       <CardContent className="space-y-6">
-        <div className="space-y-3">
-          <Label htmlFor="missionStatement" className="text-sm font-semibold text-gray-700 dark:text-gray-300">
-            Mission Statement *
-          </Label>
-          <Textarea
-            id="missionStatement"
-            name="missionStatement"
-            placeholder="Describe your organization's mission and goals..."
-            required
-            value={formData.missionStatement}
-            onChange={onInputChange("missionStatement")}
-            className={errors?.missionStatement ? "border-red-500" : ""}
-          />
-          {errors?.missionStatement && (
-            <p className="text-sm text-red-500">{errors.missionStatement[0]}</p>
+        <FormField
+          control={form.control}
+          name="missionStatement"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Mission Statement *
+              </FormLabel>
+              <FormControl>
+                <Textarea
+                  {...field}
+                  placeholder="Describe your organization's mission and goals..."
+                  className={
+                    form.formState.errors?.missionStatement
+                      ? "border-red-500"
+                      : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
           )}
-        </div>
+        />
 
         <div className="space-y-4">
           <div>
@@ -64,7 +82,9 @@ export function MissionProjectsSection({
               <div key={area} className="flex items-center space-x-2">
                 <Checkbox
                   id={area}
-                  checked={formData.projectAreas?.includes(area) || false}
+                  checked={
+                    form.getValues("projectAreas")?.includes(area) || false
+                  }
                   onCheckedChange={(checked) =>
                     onProjectAreaChange(area, checked as boolean)
                   }
@@ -75,115 +95,127 @@ export function MissionProjectsSection({
               </div>
             ))}
           </div>
-          {formData.projectAreas?.map((area) => (
-            <input key={area} type="hidden" name="projectAreas" value={area} />
-          ))}
-          {errors?.projectAreas && (
-            <p className="text-sm text-red-500">{errors.projectAreas[0]}</p>
-          )}
-        </div>
-
-        <div className="space-y-3">
-          <Label htmlFor="websiteUrl" className="text-sm font-semibold text-gray-700 dark:text-gray-300">
-            Website URL <span className="text-gray-400">(Optional)</span>
-          </Label>
-          <Input
-            id="websiteUrl"
-            name="websiteUrl"
-            type="url"
-            placeholder="https://www.yourorganization.org"
-            value={formData.websiteUrl}
-            onChange={onInputChange("websiteUrl")}
-            className={errors?.websiteUrl ? "border-red-500" : ""}
-          />
-          {errors?.websiteUrl && (
-            <p className="text-sm text-red-500">{errors.websiteUrl[0]}</p>
-          )}
-        </div>
-
-        <div className="space-y-4">
-          <div>
-            <Label className="text-sm font-semibold text-gray-700 dark:text-gray-300">
-              Social Media Links <span className="text-gray-400">(Optional)</span>
-            </Label>
-            <p className="text-xs text-muted-foreground mt-1">
-              Add your organization&apos;s social media presence
+          {form.formState.errors?.projectAreas && (
+            <p className="text-sm text-red-500">
+              {form.formState.errors.projectAreas.message}
             </p>
-          </div>
-          <div className="space-y-4 pl-4 border-l-2 border-gray-200 dark:border-gray-700">
-            <div className="space-y-2">
-              <Label htmlFor="facebookUrl" className="text-sm font-medium text-gray-600 dark:text-gray-400">
-                Facebook
-              </Label>
-              <Input
-                id="facebookUrl"
-                name="facebookUrl"
-                type="url"
-                placeholder="https://www.facebook.com/yourorganization"
-                value={formData.facebookUrl}
-                onChange={onInputChange("facebookUrl")}
-                className={errors?.facebookUrl ? "border-red-500" : ""}
-              />
-              {errors?.facebookUrl && (
-                <p className="text-sm text-red-500">{errors.facebookUrl[0]}</p>
-              )}
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="twitterUrl" className="text-sm font-medium text-gray-600 dark:text-gray-400">
-                Twitter/X
-              </Label>
-              <Input
-                id="twitterUrl"
-                name="twitterUrl"
-                type="url"
-                placeholder="https://twitter.com/yourorganization"
-                value={formData.twitterUrl}
-                onChange={onInputChange("twitterUrl")}
-                className={errors?.twitterUrl ? "border-red-500" : ""}
-              />
-              {errors?.twitterUrl && (
-                <p className="text-sm text-red-500">{errors.twitterUrl[0]}</p>
-              )}
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="instagramUrl" className="text-sm font-medium text-gray-600 dark:text-gray-400">
-                Instagram
-              </Label>
-              <Input
-                id="instagramUrl"
-                name="instagramUrl"
-                type="url"
-                placeholder="https://www.instagram.com/yourorganization"
-                value={formData.instagramUrl}
-                onChange={onInputChange("instagramUrl")}
-                className={errors?.instagramUrl ? "border-red-500" : ""}
-              />
-              {errors?.instagramUrl && (
-                <p className="text-sm text-red-500">{errors.instagramUrl[0]}</p>
-              )}
-            </div>
-            
-            <div className="space-y-2">
-              <Label htmlFor="linkedinUrl" className="text-sm font-medium text-gray-600 dark:text-gray-400">
-                LinkedIn
-              </Label>
-              <Input
-                id="linkedinUrl"
-                name="linkedinUrl"
-                type="url"
-                placeholder="https://www.linkedin.com/company/yourorganization"
-                value={formData.linkedinUrl}
-                onChange={onInputChange("linkedinUrl")}
-                className={errors?.linkedinUrl ? "border-red-500" : ""}
-              />
-              {errors?.linkedinUrl && (
-                <p className="text-sm text-red-500">{errors.linkedinUrl[0]}</p>
-              )}
-            </div>
-          </div>
+          )}
         </div>
+
+        <FormField
+          control={form.control}
+          name="websiteUrl"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Website URL
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="url"
+                  placeholder="https://www.yourorganization.org"
+                  className={
+                    form.formState.errors?.websiteUrl ? "border-red-500" : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="facebookUrl"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Facebook URL
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="url"
+                  placeholder="https://www.facebook.com/yourorganization"
+                  className={
+                    form.formState.errors?.facebookUrl ? "border-red-500" : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="twitterUrl"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Twitter URL
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="url"
+                  placeholder="https://twitter.com/yourorganization"
+                  className={
+                    form.formState.errors?.twitterUrl ? "border-red-500" : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="instagramUrl"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                Instagram URL
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="url"
+                  placeholder="https://www.instagram.com/yourorganization"
+                  className={
+                    form.formState.errors?.instagramUrl ? "border-red-500" : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="linkedinUrl"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="text-sm font-semibold text-gray-700 dark:text-gray-300">
+                LinkedIn URL
+              </FormLabel>
+              <FormControl>
+                <Input
+                  {...field}
+                  type="url"
+                  placeholder="https://www.linkedin.com/company/yourorganization"
+                  className={
+                    form.formState.errors?.linkedinUrl ? "border-red-500" : ""
+                  }
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
       </CardContent>
     </Card>
   );
